+++ JSM.Insight/Public/Get-InsightWorkspaceID.ps1
@@ -2,7 +2,7 @@
 function Get-InsightWorkspaceID {
     [CmdletBinding()]
     param (
-        [uri]$InsightServerUrl,
+        [string]$InsightServerUrl,
         [string]$InsightCreds = $InsightCreds
     )
 
@@ -18,13 +18,14 @@ function Get-InsightWorkspaceID {
     end {
         try {
             $response = Invoke-RestMethod -Uri $Request.Uri -Headers $headers -Method GET
+
+            $script:InsightWorkspaceID = $response.values.workspaceId
+            $response.values.workspaceId
         }
         catch {
             Write-Verbose "[$($MyInvocation.MyCommand.Name)] Failed"
         } 
-        $script:InsightWorkspaceID = $response.values.workspaceId
-        $response.values.workspaceId
-
+        
         Write-Verbose "[$($MyInvocation.MyCommand.Name)] Complete"
     }
 }